requirements.txt
streamlit
PyPDF2
langchain-google-genai
langchain
langchain-core
python-dotenv
radon
bandit
